Plot Break Down Objects in D3 with r2d3

Plot Break Down Objects in D3 with r2d3

plotD3(x, ...)

# S3 method for break_down
plotD3(x, ..., max_features = 4, min_max = NA,
  vcolors = c(`-1` = "#a3142f", `0` = "#a3142f", `1` = "#0f6333", X =
  "#0f6333"), digits = 3, rounding_function = round)

Arguments

x
the model model of 'break_down' class

...
other parameters

max_features
default 4. Maximal number of features to be included in the plot

min_max
range of OX axis. By deafult `NA` therefore will be extracted from the contributions of `x`. But can be set to some constants, usefull if these plots are used for comparisons.

vcolors
named vector with colors

digits
number of decimal places (round) or significant digits (signif) to be used. See the rounding_function argument

rounding_function
function that is to used for rounding numbers. It may be signif() which keeps a specified number of significant digits. Or the default round() to have the same precision for all components

Value

a r2d3 object

Examples

## Not run:
# prepare dataset
library("titanic")
titanic <- titanic_train[,c("Survived", "Pclass", "Sex", "Age",
                            "SibSp", "Parch", "Fare", "Embarked")]
titanic$Survived <- factor(titanic$Survived)
titanic$Sex <- factor(titanic$Sex)
titanic$Embarked <- factor(titanic$Embarked)
titanic <- na.omit(titanic)

# prepare model
library("randomForest")
rf_model <- randomForest(Survived ~ .,  data = titanic)

# prepare explainer
library("DALEX2")
predict_fuction <- function(m,x) predict(m, x, type = "prob")[,2]
rf_explain <- explain(rf_model, data = titanic,
                      y = titanic$Survived == "1", label = "RF",
                      predict_function = predict_fuction)

# plor D3 explainers
library("breakDown2")
rf_la <- local_attributions(rf_explain, titanic[2,])
rf_la

#>                    contribution
#> RF: baseline              0.363
#> RF: * Sex = female        0.302
#> RF: * Pclass = 1          0.181
#> RF: * Fare = 71           0.101
#> RF: * Embarked = C        0.015
#> RF: * Age = 38            0.019
#> RF: * SibSp = 1           0.008
#> RF: * Parch = 0           0.006
#> RF: * Survived = 1        0.000
#> RF: prediction            0.996
#> baseline:  0 

plotD3(rf_la)

rf_la <- local_attributions(rf_explain, titanic[3,])
rf_la

#>                    contribution
#> RF: baseline              0.363
#> RF: * Sex = female        0.302
#> RF: * Pclass = 3         -0.220
#> RF: * Fare = 7.9         -0.049
#> RF: * SibSp = 0           0.035
#> RF: * Embarked = S       -0.022
#> RF: * Age = 26            0.137
#> RF: * Parch = 0          -0.040
#> RF: * Survived = 1        0.000
#> RF: prediction            0.506
#> baseline:  0 

plotD3(rf_la, max_features = 10)
plotD3(rf_la, max_features = 10, min_max = c(0,1))